+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ ADDED:
     `chaospy` to generate quasi-random samples and orthogonal polynomials.
 CHANGED:
   * `chaospy.Trunc` updated to take both `lower` and `upper` at the same time.
+REMOVED:
+  * `chaospy.SampleDist` removed in favor of `chaospy.GaussianKDE`.
 
 Version 4.0-beta3 (2020-10-22)
 ==============================